requirements.txt
fastapi==0.85.0
alembic==1.8.1
uvicorn==0.18.3
email-validator==1.3.0
python-decouple==3.6
psycopg2-binary==2.9.3
requests==2.28.1
pydantic_sqlalchemy==0.0.9
